feat: add workflow step to set cluster info as environment variables
The workflow file was updated to include a new step that sets the cluster name and zone as environment variables. This step uses the `terraform output` command to retrieve the values and then uses the `echo` command to set the outputs. This will allow other steps in the workflow to access and use this information.